fix: 마이페이지 내 state가 업데이트되면서 api가 계속 호출되는 이슈 핸들링 및 컴포넌트 리팩토링
